Removed billing info from edit profile page. Added billing info modal before plan upgrading.
